restructuring buffers into inputs
